@@ -4,7 +4,7 @@ import { AppRoutingModule } from './app-routing.module'
 import { AppComponent } from './app.component'
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations'
 import { MatButtonModule } from '@angular/material/button'
-import { HttpClientModule } from '@angular/common/http'
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http'
 import { ReactiveFormsModule } from '@angular/forms'
 import { FormlyModule } from '@ngx-formly/core'
 import { FormlyMaterialModule } from '@ngx-formly/material'
@@ -20,14 +20,22 @@ import { MatCheckboxModule } from '@angular/material/checkbox'
 import { MatNativeDateModule } from '@angular/material/core'
 import { MatTableModule } from '@angular/material/table';
 import {MatListModule} from '@angular/material/list';
+import {MatPaginatorModule} from '@angular/material/paginator';
 import { EmploeeFormComponent, EmploeeListComponent } from './emploee'
 import { AppConfig } from './app.config'
+import { HomeComponent } from './home/home.component'
+import { NavMenuComponent } from './nav-menu/nav-menu.component'
+import { LoginComponent } from './login/login.component'
+import { ErrorInterceptor, JwtInterceptor } from './interceptors'
 
 @NgModule({
     declarations: [
         AppComponent,
         EmploeeListComponent,
         EmploeeFormComponent,
+        HomeComponent,
+        NavMenuComponent,
+        LoginComponent
     ],
     imports: [
         BrowserModule,
@@ -52,7 +60,10 @@ import { AppConfig } from './app.config'
         FormlyMatDatepickerModule,
         FormlyMatToggleModule,
     ],
-    providers: [],
+    providers: [
+        { provide: HTTP_INTERCEPTORS, useClass: JwtInterceptor, multi: true },
+        { provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true },
+    ],
     bootstrap: [AppComponent],
 })
 export class AppModule {}
