@@ -3,7 +3,6 @@ import { FormGroup } from '@angular/forms';
 import { FormlyFieldConfig, FormlyFormOptions } from '@ngx-formly/core';
 import { BehaviorSubject, Observable } from 'rxjs';
 
-@Injectable({ providedIn: 'root' })
 export class FormService {
   public form: FormGroup = new FormGroup({}, { updateOn: 'blur' });
   public model: any = {};
@@ -23,7 +22,7 @@ export class FormService {
     Object.entries(errors).forEach(([key, value]: any) => {
       const field = this.fix(key);
       console.log(key, value, this.form, this.form.get(field));
-      const error = value[0]?.replace(/\W+/, '');
+      const error = (value instanceof Array && value.length > 0 ? value[0] : value);
       this.form.get(field)?.setErrors({ server: { message: error }});
     });
   }
