@@ -1,38 +1,28 @@
 import { Injectable } from '@angular/core';
-import {
-  HttpInterceptor,
+import { 
   HttpRequest,
-  HttpErrorResponse,
-  HttpHandler,
-  HttpEvent
+  HttpHandler, 
+  HttpEvent, 
+  HttpInterceptor 
 } from '@angular/common/http';
-
-import { Observable, EMPTY } from 'rxjs';
+import { Observable, throwError } from 'rxjs';
 import { catchError } from 'rxjs/operators';
+import { Router } from '@angular/router';
+import { JwtService } from '../login';
 
 @Injectable()
 export class ErrorInterceptor implements HttpInterceptor {
-  intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
+    constructor(private jwtService: JwtService, private router: Router) { console.log('ErrorInterceptor'); }
 
-    return next.handle(request).pipe(catchError((error: HttpErrorResponse) => {
-        if (error.error instanceof Error) {
-          // A client-side or network error occurred. Handle it accordingly.
-          console.error('An error occurred:', error.error.message);
-        } else {
-          // The backend returned an unsuccessful response code.
-          // The response body may contain clues as to what went wrong,
-          console.error(`Backend returned code ${error.status}, body was: ${error.error}`);
-        }
-
-        // If you want to return a new response:
-        //return of(new HttpResponse({body: [{name: "Default value..."}]}));
-
-        // If you want to return the error on the upper level:
-        //return throwError(error);
-
-        // or just return nothing:
-        return EMPTY;
-      })
-    );
-  }
+    intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
+        return next.handle(request).pipe(catchError(err => {
+            if (err.status === 401) {                
+                this.jwtService.Logout();
+                this.router.navigate(['login']);
+                return throwError('401 redirect');
+            }
+            const error = err.error.message || err.statusText;
+            return throwError(error);
+        }))
+    }
 }
